refactor: use macros for VlenV2 based codecs
