+++ zarrs/src/array/codec/array_to_bytes/vlen_array/vlen_array_codec.rs
@@ -1,132 +1,3 @@
-use std::sync::Arc;
+use crate::array::codec::array_to_bytes::vlen_v2::vlen_v2_impl;
 
-use zarrs_metadata::v3::MetadataV3;
-
-use crate::array::{
-    codec::{
-        array_to_bytes::vlen_v2::VlenV2Codec, ArrayPartialDecoderTraits, ArrayPartialEncoderTraits,
-        ArrayToBytesCodecTraits, BytesPartialDecoderTraits, BytesPartialEncoderTraits, CodecError,
-        CodecOptions, CodecTraits,
-    },
-    ArrayBytes, ArrayCodecTraits, ArrayMetadataOptions, BytesRepresentation, ChunkRepresentation,
-    RawBytes, RecommendedConcurrency,
-};
-
-#[cfg(feature = "async")]
-use crate::array::codec::{AsyncArrayPartialDecoderTraits, AsyncBytesPartialDecoderTraits};
-
-/// The `vlen-array` codec implementation.
-#[derive(Debug, Clone)]
-pub struct VlenArrayCodec {
-    inner: Arc<VlenV2Codec>,
-}
-
-impl VlenArrayCodec {
-    /// Create a new `vlen-array` codec.
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            inner: Arc::new(VlenV2Codec::new("vlen-array".to_string())),
-        }
-    }
-}
-
-impl Default for VlenArrayCodec {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl CodecTraits for VlenArrayCodec {
-    fn create_metadata_opt(&self, options: &ArrayMetadataOptions) -> Option<MetadataV3> {
-        self.inner.create_metadata_opt(options)
-    }
-
-    fn partial_decoder_should_cache_input(&self) -> bool {
-        self.inner.partial_decoder_should_cache_input()
-    }
-
-    fn partial_decoder_decodes_all(&self) -> bool {
-        self.inner.partial_decoder_decodes_all()
-    }
-}
-
-impl ArrayCodecTraits for VlenArrayCodec {
-    fn recommended_concurrency(
-        &self,
-        decoded_representation: &ChunkRepresentation,
-    ) -> Result<RecommendedConcurrency, CodecError> {
-        self.inner.recommended_concurrency(decoded_representation)
-    }
-}
-
-#[cfg_attr(feature = "async", async_trait::async_trait)]
-impl ArrayToBytesCodecTraits for VlenArrayCodec {
-    fn dynamic(self: Arc<Self>) -> Arc<dyn ArrayToBytesCodecTraits> {
-        self as Arc<dyn ArrayToBytesCodecTraits>
-    }
-
-    fn encode<'a>(
-        &self,
-        bytes: ArrayBytes<'a>,
-        decoded_representation: &ChunkRepresentation,
-        options: &CodecOptions,
-    ) -> Result<RawBytes<'a>, CodecError> {
-        self.inner.encode(bytes, decoded_representation, options)
-    }
-
-    fn decode<'a>(
-        &self,
-        bytes: RawBytes<'a>,
-        decoded_representation: &ChunkRepresentation,
-        options: &CodecOptions,
-    ) -> Result<ArrayBytes<'a>, CodecError> {
-        self.inner.decode(bytes, decoded_representation, options)
-    }
-
-    fn partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn BytesPartialDecoderTraits>,
-        decoded_representation: &ChunkRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn ArrayPartialDecoderTraits>, CodecError> {
-        self.inner
-            .clone()
-            .partial_decoder(input_handle, decoded_representation, options)
-    }
-
-    fn partial_encoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn BytesPartialDecoderTraits>,
-        output_handle: Arc<dyn BytesPartialEncoderTraits>,
-        decoded_representation: &ChunkRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn ArrayPartialEncoderTraits>, CodecError> {
-        self.inner.clone().partial_encoder(
-            input_handle,
-            output_handle,
-            decoded_representation,
-            options,
-        )
-    }
-
-    #[cfg(feature = "async")]
-    async fn async_partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn AsyncBytesPartialDecoderTraits>,
-        decoded_representation: &ChunkRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncArrayPartialDecoderTraits>, CodecError> {
-        self.inner
-            .clone()
-            .async_partial_decoder(input_handle, decoded_representation, options)
-            .await
-    }
-
-    fn compute_encoded_size(
-        &self,
-        decoded_representation: &ChunkRepresentation,
-    ) -> Result<BytesRepresentation, CodecError> {
-        self.inner.compute_encoded_size(decoded_representation)
-    }
-}
+vlen_v2_impl::vlen_v2_impl!(VlenArrayCodec, "vlen-array");
