+++ zarrs/src/array/codec/array_to_bytes/vlen_v2.rs
@@ -3,6 +3,8 @@
 mod vlen_v2_codec;
 mod vlen_v2_partial_decoder;
 
+pub(crate) mod vlen_v2_impl;
+
 use std::{mem::size_of, sync::Arc};
 
 /// The identifier for the `vlen_v2` codec.
